+++ vimiv/gui/commandline.py
@@ -70,7 +70,9 @@ class CommandLine(KeyHandler, QLineEdit):
     def __init__(self):
         super().__init__()
         self._history = history.History(
-            history.read(), max_items=api.settings.command.history_limit.value
+            self.PREFIXES,
+            history.read(),
+            max_items=api.settings.command.history_limit.value,
         )
         self.mode = None